Cast width/height variables as needed
Width & height are stored as different types depending on context. So cast them
to the correct type (whether it be float, uint32_t, int32_t, int, etc) to do
the required computation and cast again if necessary.

+++ framework/application/wayland_window.cpp
@@ -299,7 +299,7 @@ void WaylandWindow::UpdateWindowSize()
             scale_ = output_info.scale;
         }
 
-        if (output_info.width == width_ && output_info.height == height_)
+        if (output_info.width == static_cast<int32_t>(width_) && output_info.height == static_cast<int32_t>(height_))
         {
             if (xdg_toplevel_ != nullptr)
             {
